2010-03-21  Tatsuhiro Tsujikawa  <t-tujikawa@users.sourceforge.net>
	Defined cuid_t as int64_t. Removed texts containing "CUID#%D" from
	translatable texts because it is used in log and debugging
	purpose.

+++ src/CUIDCounter.h
@@ -50,7 +50,7 @@ public:
 
   cuid_t newID()
   {
-    if(_count == INT32_MAX) {
+    if(_count == INT64_MAX) {
       _count = 0;
     }
     return ++_count;
